Refactor member widgets into a separate feature section for clarity
