+++ native/acter/src/api/room.rs
@@ -115,6 +115,7 @@ impl From<MessageLikeEventType> for PermissionTest {
 }
 
 pub struct Member {
+    pub(crate) room: Room,
     pub(crate) member: RoomMember,
     pub(crate) acter_app_settings: Option<ActerAppSettingsContent>,
 }
@@ -134,6 +135,10 @@ impl Member {
 
     pub fn user_id(&self) -> OwnedUserId {
         self.member.user_id().to_owned()
+    }
+
+    pub fn room_id_str(&self) -> String {
+        self.room.room_id().to_string()
     }
 
     pub fn can_string(&self, input: String) -> bool {
@@ -288,6 +293,42 @@ impl Member {
         RUNTIME
             .spawn(async move {
                 member.unignore().await?;
+                Ok(true)
+            })
+            .await?
+    }
+
+    pub async fn kick(&self, msg: Option<String>) -> Result<bool> {
+        let member_id = self.member.user_id().to_owned();
+        let room = self.room.clone();
+
+        RUNTIME
+            .spawn(async move {
+                room.kick_user(&member_id, msg.as_deref()).await?;
+                Ok(true)
+            })
+            .await?
+    }
+
+    pub async fn ban(&self, msg: Option<String>) -> Result<bool> {
+        let member_id = self.member.user_id().to_owned();
+        let room = self.room.clone();
+
+        RUNTIME
+            .spawn(async move {
+                room.ban_user(&member_id, msg.as_deref()).await?;
+                Ok(true)
+            })
+            .await?
+    }
+
+    pub async fn unban(&self, msg: Option<String>) -> Result<bool> {
+        let member_id = self.member.user_id().to_owned();
+        let room = self.room.clone();
+
+        RUNTIME
+            .spawn(async move {
+                room.unban_user(&member_id, msg.as_deref()).await?;
                 Ok(true)
             })
             .await?
@@ -615,9 +656,9 @@ impl Room {
         if !self.is_joined() {
             bail!("Not a room we have joined");
         }
-        let room = self.room.clone();
+        let me = self.clone();
 
-        let client = room.client();
+        let client = me.room.client();
         let my_id = client
             .user_id()
             .context("You must be logged in to do that")?
@@ -631,12 +672,14 @@ impl Room {
 
         RUNTIME
             .spawn(async move {
-                let member = room
+                let member = me
+                    .room
                     .get_member(&my_id)
                     .await?
                     .context("Unable to find me in room")?;
                 Ok(Member {
                     member,
+                    room: me.clone(),
                     acter_app_settings,
                 })
             })
@@ -768,7 +811,7 @@ impl Room {
     }
 
     pub async fn active_members(&self) -> Result<Vec<Member>> {
-        let room = self.room.clone();
+        let room = self.clone();
 
         let is_acter_space = self.is_acter_space().await?;
         let acter_app_settings = if is_acter_space {
@@ -785,6 +828,7 @@ impl Room {
                     .into_iter()
                     .map(|member| Member {
                         member,
+                        room: room.clone(),
                         acter_app_settings: acter_app_settings.clone(),
                     })
                     .collect();
@@ -817,7 +861,7 @@ impl Room {
     }
 
     pub async fn invited_members(&self) -> Result<Vec<Member>> {
-        let room = self.room.clone();
+        let room = self.clone();
         let is_acter_space = self.is_acter_space().await?;
         let acter_app_settings = if is_acter_space {
             Some(self.app_settings_content().await?)
@@ -833,6 +877,7 @@ impl Room {
                     .into_iter()
                     .map(|member| Member {
                         member,
+                        room: room.clone(),
                         acter_app_settings: acter_app_settings.clone(),
                     })
                     .collect();
@@ -842,7 +887,7 @@ impl Room {
     }
 
     pub async fn active_members_no_sync(&self) -> Result<Vec<Member>> {
-        let room = self.room.clone();
+        let room = self.clone();
         let is_acter_space = self.is_acter_space().await?;
         let acter_app_settings = if is_acter_space {
             Some(self.app_settings_content().await?)
@@ -858,6 +903,7 @@ impl Room {
                     .into_iter()
                     .map(|member| Member {
                         member,
+                        room: room.clone(),
                         acter_app_settings: acter_app_settings.clone(),
                     })
                     .collect();
@@ -868,6 +914,7 @@ impl Room {
 
     pub async fn get_member(&self, user_id: String) -> Result<Member> {
         let room = self.room.clone();
+        let api_room = self.clone();
         let uid = UserId::parse(user_id)?;
         let is_acter_space = self.is_acter_space().await?;
         let acter_app_settings = if is_acter_space {
@@ -884,6 +931,7 @@ impl Room {
                     .context("Unable to find user in room")?;
                 Ok(Member {
                     member,
+                    room: api_room.clone(),
                     acter_app_settings: acter_app_settings.clone(),
                 })
             })
@@ -1167,6 +1215,7 @@ impl Room {
             bail!("Unable to get a room we are not invited");
         }
         let room = self.room.clone();
+        let api_room = self.clone();
         let is_acter_space = self.is_acter_space().await?;
         let acter_app_settings = if is_acter_space {
             Some(self.app_settings_content().await?)
@@ -1185,6 +1234,7 @@ impl Room {
                     if let Some(member) = room.get_member(user_id).await? {
                         members.push(Member {
                             member,
+                            room: api_room.clone(),
                             acter_app_settings: acter_app_settings.clone(),
                         });
                     }
